Wired up iframe loader for preview
